% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmode.R
\name{getmode}
\alias{getmode}
\title{Find Mode}
\usage{
getmode(v)
}
\arguments{
\item{v}{vector containing points}
}
\value{
The mode of the vector
}
\description{
Find Mode
}
\examples{
v = c(1, 4, 4, 4, 5, 6)
getmode(v)
}
