% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myhyper.R
\name{myhyper}
\alias{myhyper}
\title{Hypergeometric Simulation}
\usage{
myhyper(iter = 100, N = 20, r = 12, n = 5)
}
\arguments{
\item{iter}{the number of iterations}

\item{N}{the size of the population}

\item{r}{the number of successes}

\item{n}{the sample size}
}
\value{
Returns a barplot and table for a hypergeometric simulation
}
\description{
Hypergeometric Simulation
}
